Update to 3.6+ only
