+++ .travis.yml
@@ -2,13 +2,9 @@ sudo: required
 dist: trusty
 language: python
 python:
-    - "3.5"
     - "3.6"
-matrix:
-  include:
-    - python: "3.7"
-      dist: xenial
-      sudo: true
+    - "3.7"
+    - "3.8-dev"
 install:
     - pip install -r requirements.txt
     - pip install -r requirements-test.txt
